rename doc_init to _doc and fix pep8
