+++ calour/_doc.py
@@ -1,4 +1,3 @@
 import docrep
 
 ds = docrep.DocstringProcessor()
-
